Add ExMemberPage
We're adding this class as we will eventually want to scrape
the end dates of members who have left the current legislature.

The class will have the exact same behaviour as its parent class
apart from #electoral_history, which will return a hash populated
with dates from the members electoral history. (For a future PR)

This commit is a step towards that.